Conflicts:
	ganttzk/src/main/java/org/zkoss/ganttz/TaskList.java
	ganttzk/src/main/resources/web/ganttz/zul/plannerLayout.zul
	navalplanner-webapp/src/main/java/org/navalplanner/web/planner/tabs/MultipleTabsPlannerController.java

FEA: ItEr75S08MigrationZK5

+++ navalplanner-webapp/src/main/java/org/navalplanner/web/planner/allocation/AllocationRow.java
@@ -419,7 +419,7 @@ public abstract class AllocationRow {
 
     private void setUnknownResourcesPerDay() {
         this.editedValue = null;
-        this.intendedResourcesPerDayInput.setValue(null);
+        this.intendedResourcesPerDayInput.setValue((BigDecimal) null);
         clearRealResourcesPerDay();
     }
